Fix bug
ValueError was raised inside the try block and because the except block
checked for ValueError, the exception was never passed to the caller.

Add test to check if it is working. The test passes even without the bug
fix because a ValueError is raised when linking the item code, but this
is not the correct behaviour